requirements.txt
# Core AI and RAG Orchestration
langchain==0.3.21
langchain-core==0.3.46
langchain-community==0.3.20
torch==2.6.0
sentence-transformers==2.2.2

# Vector Search
faiss-cpu==1.10.0

# Document Parsers
pymupdf==1.25.4
python-docx==0.8.11
python-pptx==0.6.21
openpyxl==3.1.2

# Configuration
python-dotenv==1.0.1
